Runs without deadlocks now. Gossip updates are rejected, but closer to functional
