+++ Cargo.toml
@@ -44,12 +44,12 @@ tower-http = { version = "0.6.6", features = [
 ] }
 tower-layer = "0.3.2"
 tower-service = "0.3.2"
-thiserror = "1.0"
+thiserror = "2.0.17"
 url = "2.5.7"
 tracing = "0.1.41"
 tracing-futures = "0.2.5"
 tracing-subscriber = { version = "0.3.20", features = ["env-filter", "json"] }
-
+console-subscriber = "0.4.1"
 
 [dev-dependencies]
 criterion = "0.7"
