+++ src/node/gossip/mod.rs
@@ -4,11 +4,13 @@
 //! versioned token buckets with vector clocks for conflict resolution,
 //! transport mechanisms, and scheduling logic to efficiently propagate
 //! rate limit state among nodes.
+pub mod controller;
 pub mod gossip_node;
 pub mod messages;
 
+pub use controller::GossipController;
 pub use gossip_node::GossipNode;
-pub use messages::{GossipMessage, GossipPacket};
+pub use messages::{GossipCommand, GossipMessage, GossipPacket};
 
 // Additional modules will be added as we implement them:
 // pub mod membership_manager;
